extract ConversationUpdateRole
